config: Fix kconfig warnings
We're getting the following warnings:

```
 #13 4.361 .config:85:warning: override: reassigning to symbol DEBUG_INFO_BTF
 #13 4.361 .config:86:warning: override: DEBUG_INFO_DWARF_TOOLCHAIN_DEFAULT changes choice state
 #13 4.361 .config:208:warning: override: reassigning to symbol VSOCKETS
```

The first and last are legit - we are double assigning. So just delete
the second assignments.

The middle we want to keep. It's probably safer to use the latest DWARF
version possible - pahole will be able to handle it. Older versions, who
knows. Safer to stay on bleeding edge, as it's what BPF subsystem tests
with.